refactor(web): remove app.css and move theme to its own folder
